:sparkles: Add Firefox profile support

+++ src/webext/dbus.rs
@@ -3,29 +3,29 @@ use dbus::arg;
 use dbus::blocking;
 
 pub trait DevPustaczekVaxtify {
-	fn browser_register(&self, pid: i64) -> Result<(), dbus::Error>;
-	fn browser_tab_delete(&self, pid: i64, tab: i64) -> Result<(), dbus::Error>;
-	fn browser_tab_update(&self, pid: i64, tab: i64, url: &str) -> Result<(), dbus::Error>;
-	fn browser_unregister(&self, pid: i64) -> Result<(), dbus::Error>;
+	fn browser_register(&self, pid: u32) -> Result<(), dbus::Error>;
+	fn browser_tab_delete(&self, pid: u32, tab: i32) -> Result<(), dbus::Error>;
+	fn browser_tab_update(&self, pid: u32, tab: i32, url: &str) -> Result<(), dbus::Error>;
+	fn browser_unregister(&self, pid: u32) -> Result<(), dbus::Error>;
 	fn permit_end(&self, permit: &str) -> Result<(), dbus::Error>;
 	fn permit_start(&self, permit: &str) -> Result<(), dbus::Error>;
 	fn permit_start_with_duration(&self, permit: &str, duration: u64) -> Result<(), dbus::Error>;
 }
 
 impl<'a, T: blocking::BlockingSender, C: ::std::ops::Deref<Target = T>> DevPustaczekVaxtify for blocking::Proxy<'a, C> {
-	fn browser_register(&self, pid: i64) -> Result<(), dbus::Error> {
+	fn browser_register(&self, pid: u32) -> Result<(), dbus::Error> {
 		self.method_call("dev.pustaczek.Vaxtify", "BrowserRegister", (pid,))
 	}
 
-	fn browser_tab_delete(&self, pid: i64, tab: i64) -> Result<(), dbus::Error> {
+	fn browser_tab_delete(&self, pid: u32, tab: i32) -> Result<(), dbus::Error> {
 		self.method_call("dev.pustaczek.Vaxtify", "BrowserTabDelete", (pid, tab))
 	}
 
-	fn browser_tab_update(&self, pid: i64, tab: i64, url: &str) -> Result<(), dbus::Error> {
+	fn browser_tab_update(&self, pid: u32, tab: i32, url: &str) -> Result<(), dbus::Error> {
 		self.method_call("dev.pustaczek.Vaxtify", "BrowserTabUpdate", (pid, tab, url))
 	}
 
-	fn browser_unregister(&self, pid: i64) -> Result<(), dbus::Error> {
+	fn browser_unregister(&self, pid: u32) -> Result<(), dbus::Error> {
 		self.method_call("dev.pustaczek.Vaxtify", "BrowserUnregister", (pid,))
 	}
 
@@ -44,8 +44,8 @@ impl<'a, T: blocking::BlockingSender, C: ::std::ops::Deref<Target = T>> DevPusta
 
 #[derive(Debug)]
 pub struct DevPustaczekVaxtifyTabClose {
-	pub pid: i64,
-	pub tab: i64,
+	pub pid: u32,
+	pub tab: i32,
 }
 
 impl arg::AppendAll for DevPustaczekVaxtifyTabClose {
@@ -68,7 +68,7 @@ impl dbus::message::SignalArgs for DevPustaczekVaxtifyTabClose {
 
 #[derive(Debug)]
 pub struct DevPustaczekVaxtifyTabCreateEmpty {
-	pub pid: i64,
+	pub pid: u32,
 }
 
 impl arg::AppendAll for DevPustaczekVaxtifyTabCreateEmpty {
